+++ packages/stream_video/lib/src/webrtc/media/camera_constraints.dart
@@ -4,16 +4,19 @@ import 'constraints/facing_mode.dart';
 import 'video_constraints.dart';
 
 /// Options used when creating a video track that captures the camera.
+/// [mirror]: Whether the video should be mirrored. If set to null, the default is true when facingMode is user and false when facingMode is environment.
 class CameraConstraints extends VideoConstraints {
   const CameraConstraints({
-    super.deviceId,
     this.facingMode = FacingMode.user,
+    this.mirror,
+    super.deviceId,
     super.maxFrameRate,
     super.params = RtcVideoParametersPresets.h720_16x9,
   });
 
   CameraConstraints.from({required VideoConstraints constraints})
       : facingMode = FacingMode.user,
+        mirror = null,
         super(
           deviceId: constraints.deviceId,
           maxFrameRate: constraints.maxFrameRate,
@@ -21,12 +24,14 @@ class CameraConstraints extends VideoConstraints {
         );
 
   final FacingMode facingMode;
+  final bool? mirror;
 
   @override
   Map<String, dynamic> toMap() {
     final constraints = <String, dynamic>{
       ...super.toMap(),
       'facingMode': facingMode.alias,
+      if (mirror != null) 'mirror': mirror,
     };
     if (deviceId != null) {
       if (CurrentPlatform.isWeb) {
@@ -52,11 +57,13 @@ class CameraConstraints extends VideoConstraints {
     FacingMode? facingMode,
     String? deviceId,
     double? maxFrameRate,
+    bool? Function()? mirror,
   }) =>
       CameraConstraints(
         params: params ?? this.params,
         facingMode: facingMode ?? this.facingMode,
         deviceId: deviceId ?? this.deviceId,
         maxFrameRate: maxFrameRate ?? this.maxFrameRate,
+        mirror: mirror?.call() ?? this.mirror,
       );
 }